Fix query to custom compare and extract table
