+++ app/javascript/app/pages/ndc-compare-all-targets/ndc-compare-all-targets-styles.scss
@@ -58,36 +58,8 @@
 
   .submitIcon {
     margin-right: 10px;
+    height: 30px;
+    width: 30px;
   }
 }
 
-.submitIcon {
-  height: 30px;
-  width: 30px;
-}
-
-.iconButton {
-  width: 100%;
-  height: 100%;
-  border: solid $gray2 1px;
-  border-width: 0 0 1px 1px;
-  border-radius: 0;
-
-  &.active {
-    background-color: $theme-color;
-  }
-
-  &.clickable {
-    cursor: pointer;
-  }
-
-  &.lastColumn {
-    border-right: solid $gray2 1px;
-  }
-}
-
-.loaderWrapper {
-  display: flex;
-  justify-content: center;
-  min-height: 460px;
-}
