+++ app/javascript/app/pages/ndc-compare-all-targets/ndc-compare-all-targets-table/ndc-compare-all-targets-table-styles.scss
@@ -1,0 +1,32 @@
+@import '~styles/layout.scss';
+
+.iconButton {
+  width: 100%;
+  height: 100%;
+  border: solid $gray2 1px;
+  border-width: 0 0 1px 1px;
+  border-radius: 0;
+
+  &.active {
+    background-color: $theme-color;
+  }
+
+  &.clickable {
+    cursor: pointer;
+  }
+
+  &.lastColumn {
+    border-right: solid $gray2 1px;
+  }
+}
+
+.loaderWrapper {
+  display: flex;
+  justify-content: center;
+  min-height: 460px;
+}
+
+.submitIcon {
+  height: 30px;
+  width: 30px;
+}
