+++ app/javascript/app/pages/ndc-compare-all-targets/ndc-compare-all-targets-table/ndc-compare-all-targets-table.js
@@ -1,0 +1,1 @@
+export { default } from './ndc-compare-all-targets-table-component';
